move ocean uniforms out of OceanPostProcess
